abstract common logic from most exporters out into common.CollectorIntegration
